% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.dbc.R
\name{download.dbc}
\alias{download.dbc}
\title{download.dbc: Baixa dados.dbc do site DataSUS}
\usage{
download.dbc(UF, inicio, final)
}
\arguments{
\item{UF}{uma strign, nome do estado}

\item{inicio}{um number de 6 digitos, os 4 primeiro sao do ano e os 2 ultimos sao do mes}

\item{final}{um number de 6 digitos, os 4 primeiro sao do ano e os 2 ultimos sao do mes}
}
\description{
A funcao download.dbc acessa o site do DataSUS para baixar os dados de Producao Hospitalar (SIH) e Producao Ambulatorial (SIA) de um estado especifico dentro de uma faixa de tempo determinada.
}
\examples{
download.dbc("CE",202301,202302)

}
