% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler.dbc.R
\name{ler.dbc}
\alias{ler.dbc}
\title{ler.dbc: Funcao para ler arquivos .dbc}
\usage{
ler.dbc(UF, caminho)
}
\arguments{
\item{UF}{Uma String, Nome do estado}

\item{caminho}{Uma String, caminho ate o diretorio onde foi baixado os dados dbc}
}
\description{
Esta funcao recebe o caminho da pasta contendo os dados(arquivos) que foram baixados em formato .dbc. Ela ler esses arquivos, realiza um recorte nessas bases de dados e  salva todas em um unico arquivo .Rdata
}
\examples{
ler.dbc("CE","C:/Users/gustaavobraga/Documents")

}
