% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel.data.sus.R
\name{painel.data.sus}
\alias{painel.data.sus}
\title{painel.data.sus: Funcao principal}
\usage{
painel.data.sus(UF, inicio, final, dir = ".")
}
\arguments{
\item{UF}{uma strign, nome do estado}

\item{inicio}{um number de 6 digitos, os 4 primeiro sao do ano e os 2 ultimos sao do mes}

\item{final}{um number de 6 digitos, os 4 primeiro sao do ano e os 2 ultimos sao do mes}

\item{dir}{Uma string indicando o diretório de saída onde os arquivos serão armazenados (o padrão é o diretório criado dentro da pasta Documentos do Usuário).}
}
\description{
Esta funcao permite executar todas as outras funcaes do pacote "PainelDataSUSGustavo" de uma vez so.
}
\examples{
painel.data.sus('CE',202301,202302,".")

}
