% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.painel.R
\name{script.painel}
\alias{script.painel}
\title{script.painel: Script do Painel do Power BI}
\usage{
script.painel(UF)
}
\arguments{
\item{UF}{Uma String, Nome do estado}
}
\description{
Essa funcao faz um tratamento na base de dados que vai ser usada no painel do Power BI. Essa base de dados foi criada pela funcao ler.dbc.R.
}
\examples{
script.painel("CE")

}
